clear all functionality added as a module.
