+++ src/curd.js
@@ -12,12 +12,4 @@ document.querySelector('.newItem').addEventListener('keydown', (e) => {
   if (e.code === 'Enter') gettask();
 });
 
-const clearAll = document.querySelector('.clearAll');
-clearAll.addEventListener('click', () => {
-  const filteredData = tasks.filter((e) => e.completed === false);
-  filteredData.forEach((e, i) => e.id = i + 1); // eslint-disable-line
-  localStorage.setItem('newItem', JSON.stringify(filteredData));
-  window.location.reload();
-});
-
 export default tasks;
